新エフェクト追加とコードの簡素化・最適化
- `AetheriusEnemy_Attack_2Effect` と `AetheriusEnemy_Attack_3Effect` の追加により、エネミー攻撃時のエフェクト再生を実装。
- JSON 入出力の統一的な実装を各クラスに適用。
- `EffekseerEffectManager` クラスを削除し、関連コードを整理。
- `sideDistance` プロパティを追加し、エフェクトの横方向距離を調整可能に。
- カメラ挙動やスコア表示の改善、タイマー管理のリファクタリングを実施。
- 不要なコードやメソッドを削除し、コードベースを簡素化。
- アプリ非アクティブ時の CPU 負荷軽減処理を追加。

+++ Src/Application/GameObject/HUD/HpBar/HpBar.cpp
@@ -12,7 +12,7 @@ void HpBar::Init()
 
 	// 初期は満タン表示
 	m_hpRate = 1.0f;
-	m_rect.width = static_cast<float>(1500.0f * m_hpRate);
+	m_rect.width = static_cast<long>(1500.0f * m_hpRate);
 
 	SceneManager::Instance().GetObjectWeakPtr(m_player);
 }
@@ -48,9 +48,7 @@ void HpBar::DrawSprite()
 }
 
 void HpBar::Update()
-{
-	// 目標のHP割合を算出（0～1にクランプ）
-	
+{	
 	if (auto playerStatus = m_player.lock(); playerStatus)
 	{
 
@@ -64,7 +62,7 @@ void HpBar::Update()
 		m_hpRate += targetHpRate - m_hpRate;
 
 		// 表示幅を更新
-		m_rect.width = static_cast<float>(1500.0f * m_hpRate);
+		m_rect.width = static_cast<long>(1500.0f * m_hpRate);
 
 		// 行列更新
 		m_mWorld = Math::Matrix::CreateScale(m_scale);
